Added unit tests for GenericDetailManager
